% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat.R
\name{cat0}
\alias{cat0}
\title{Equal to \code{cat(..., sep="")}}
\usage{
cat0(..., sep = "")
}
\arguments{
\item{...}{objects passed on to \code{cat}}

\item{sep}{a character vector of strings to append after each element}
}
\description{
Equal to \code{cat(..., sep="")}
}
\seealso{
Other cat: 
\code{\link{cat0n}()},
\code{\link{catn}()}
}
\concept{cat}
