% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat.R
\name{cat0n}
\alias{cat0n}
\title{Equal to \code{cat(..., "\\n", sep="")}}
\usage{
cat0n(..., sep = "")
}
\arguments{
\item{...}{objects passed on to \code{cat}}

\item{sep}{a character vector of strings to append after each element}
}
\description{
Equal to \code{cat(..., "\\n", sep="")}
}
\seealso{
Other cat: 
\code{\link{cat0}()},
\code{\link{catn}()}
}
\concept{cat}
