% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat2.R
\name{cat2}
\alias{cat2}
\alias{cat0}
\alias{catn}
\alias{cat0n}
\alias{catsn}
\alias{catnn}
\title{Concatenate and Print}
\usage{
cat2(..., sep, end)

cat0(..., sep = "", end = "")

catn(..., sep = " ", end = "\\n")

cat0n(..., sep = "", end = "\\n")

catsn(..., sep = " ", end = "\\n")

catnn(..., sep = "\\n", end = "\\n")
}
\arguments{
\item{...}{objects passed on to \link[base]{cat}}

\item{sep}{a character vector of strings to append after each element}

\item{end}{a string to print after all other elements}
}
\value{
No return value, called for side effects
}
\description{
Same as \code{cat} but with an additional argument \code{end},
which gets printed after all other elements.
}
\examples{
cat0("hello", "world") # prints "helloworld" (without newline)
catn("hello", "world") # prints "hello world\n"
cat0n("hello", "world") # prints "helloworld\n"
catsn("hello", "world") # prints "hello world\n"
catnn("hello", "world") # prints "hello\nworld\n"
}
