% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depr.R
\name{catfn}
\alias{catfn}
\title{Format and Print}
\usage{
catfn(
  fmt,
  ...,
  end = "\\n",
  file = "",
  sep = " ",
  fill = FALSE,
  labels = NULL,
  append = FALSE
)
}
\arguments{
\item{fmt}{passed on to \code{\link[base:sprintf]{base::sprintf()}}}

\item{...}{passed on to \code{\link[base:sprintf]{base::sprintf()}}}

\item{end}{passed on to \code{\link[=cat2]{cat2()}}}

\item{file}{passed on to \code{\link[=cat2]{cat2()}} (which passes it on to \code{\link[base:cat]{base::cat()}})}

\item{sep}{passed on to \code{\link[=cat2]{cat2()}} (which passes it on to \code{\link[base:cat]{base::cat()}})}

\item{fill}{passed on to \code{\link[=cat2]{cat2()}} (which passes it on to \code{\link[base:cat]{base::cat()}})}

\item{labels}{passed on to \code{\link[=cat2]{cat2()}} (which passes it on to \code{\link[base:cat]{base::cat()}})}

\item{append}{passed on to \code{\link[=cat2]{cat2()}} (which passes it on to \code{\link[base:cat]{base::cat()}})}
}
\value{
No return value, called for side effects
}
\description{
Same as \code{cat2(sprintf(fmt, ...))}

Warning: this function is deprecated and should no longer be used. The
function is guaranteed to be available as part of the package until the end
of 2023 but might removed at any time after 31.12.2023.
}
\examples{
catfn("A\%dB\%sC", 2, "asdf") # prints "A2BasdfC\n"
}
\keyword{depr}
