% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat.R
\name{catn}
\alias{catn}
\title{Equal to \code{cat(..., "\\n")}}
\usage{
catn(...)
}
\arguments{
\item{...}{objects passed on to \code{cat}}
}
\description{
Equal to \code{cat(..., "\\n")}
}
\seealso{
Other cat: 
\code{\link{cat0n}()},
\code{\link{cat0}()}
}
\concept{cat}
