% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corn.R
\name{corn}
\alias{corn}
\title{Return Corners of Matrix like Objects}
\usage{
corn(x, n = 2L)
}
\arguments{
\item{x}{matrix like object}

\item{n}{number of cols/rows from each corner to return}
}
\value{
\code{x[c(1:n, N-n:N), c(1:n, N-n:N)]}
}
\description{
Like head and tail, but returns \code{n} rows/cols from each side of
\code{x} (i.e. the corners of \code{x})
}
\examples{
corn(matrix(1:10000, 100))
}
