% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\docType{data}
\name{DOCSTRING_TEMPLATE}
\alias{DOCSTRING_TEMPLATE}
\title{Docstring Template}
\format{
A single string, i.e. a character vector of length 1.
}
\usage{
DOCSTRING_TEMPLATE
}
\description{
A minimal docstring template
}
\examples{
DOCSTRING_TEMPLATE
}
\keyword{func}
