% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{find_description_file}
\alias{find_description_file}
\title{Find DESCRIPTION File}
\usage{
find_description_file(start_dir = getwd())
}
\arguments{
\item{start_dir}{The starting directory for the search. Defaults to the current working directory.}
}
\value{
The path to the DESCRIPTION file if found. If not found, the function stops with an error message.
}
\description{
Searches for a DESCRIPTION file starting from the current or specified directory and moving upwards through the directory hierarchy until the file is found or the root directory is reached.
}
\examples{
# Start search from a specific directory
graphics_pkg_dir <- system.file(package = "graphics")
find_description_file(graphics_pkg_dir)

\dontrun{
# Below example will only work if executed from a package directory
find_description_file()
}
}
\keyword{doc}
