% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{getfd}
\alias{getfd}
\title{Get File Directory}
\usage{
getfd(
  on.error = stop("No file sourced. Maybe you're in an interactive shell?", call. =
    FALSE),
  winslash = "/"
)
}
\arguments{
\item{on.error}{Expression to use if the current file directory cannot be
determined. In that case, \code{normalizePath(on.error, winslash)} is returned.
Can also be an expression like \code{stop("message")} to stop execution
(default).}

\item{winslash}{Path separator to use for windows}
}
\value{
Current file directory as string
}
\description{
Return full path to current file directory
}
\examples{
\dontrun{
getfd()
}
getfd(on.error = getwd())
}
\keyword{path}
