% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpd.R
\name{getpd}
\alias{getpd}
\title{getpd}
\usage{
getpd(root.files = c(".git", "DESCRIPTION", "NAMESPACE"))
}
\arguments{
\item{root.files}{if any those files is found in a parent folder, that folder
is returned}
}
\value{
project root directory as string
}
\description{
get project directory
}
