% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{home}
\alias{home}
\title{Get USERPROFILE or HOME}
\usage{
home(winslash = "/")
}
\arguments{
\item{winslash}{path separator to be used on Windows (passed on to \code{normalizePath})}
}
\value{
normalized value of environment variable USERPROFILE, if defined, else value of HOME.
}
\description{
Returns normalized value of environment variable USERPROFILE, if defined, else value of HOME.
}
\examples{
home()
}
\keyword{path}
