% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifthen.R
\name{ifthen}
\alias{ifthen}
\title{Sum of vector elements}
\usage{
ifthen(...)
}
\arguments{
\item{...}{pairs of checks and corresponding return values}
}
\description{
\code{ifthen(a, b, c, d, e, f, ...)} == \code{if (a) b else if (c) d else if (e) f}
}
\examples{
x <- 2; y <- 2; z <- 1
ifthen(x==0, "foo", y==0, "bar", z==1, "this string gets returned")
}
