% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.none.R
\name{is.none}
\alias{is.none}
\title{is.none}
\usage{
is.none(x)
}
\arguments{
\item{x}{object to test}
}
\value{
TRUE or FALSE
}
\description{
TRUE for FALSE, 0, NULL, NA, empty lists and empty string
}
\examples{
is.none(FALSE) # TRUE
is.none(0) # TRUE
is.none(NA) # TRUE
is.none(list()) # TRUE
is.none("") # TRUE
is.none(1) # FALSE
}
