% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/now_ms.R
\name{now_ms}
\alias{now_ms}
\title{Get Current Date and Time as string}
\usage{
now_ms()
}
\value{
Current system time as string of the form "YYYY-MM-DD hh:mm:ss.XX
TZ", where XX means "milliseconds" and TZ means "timezone".
}
\description{
\code{now_ms} returns current system time as string of the form
"YYYY-MM-DD hh:mm:ss.XX TZ", where XX means "milliseconds" and TZ means
"timezone".
}
\examples{
now() # something like "2022-06-30, 07:14:26.82 CEST"
}
\seealso{
\code{\link[=now]{now()}}, \code{\link[=Sys.time]{Sys.time()}}, \code{\link[=format.POSIXct]{format.POSIXct()}}
}
