% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{op-null-default}
\alias{op-null-default}
\alias{\%none\%}
\title{Return Default if None}
\usage{
x \%none\% y
}
\arguments{
\item{x}{object to test}

\item{y}{object to return if \code{is.none(x)}}
}
\value{
Returns \code{y} if \code{is.none(x)} else \code{x}
}
\description{
Returns the second operand if the first operand is \code{NULL}, \code{FALSE}, \code{0}, \code{NA}, an empty list, or an empty string. Otherwise returns the first operand.
}
\examples{
FALSE \%none\% 2 # returns 2
0 \%none\% 2 # returns 2
NA \%none\% 2 # returns 2
list() \%none\% 2 # returns 2
"" \%none\% 2 # returns 2
1 \%none\% 2 # returns 1
}
\seealso{
\code{\link[=is.none]{is.none()}}
}
\keyword{check}
