% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live.R
\name{rm_all}
\alias{rm_all}
\title{Remove all objects from global environment}
\usage{
rm_all()
}
\value{
No return value, called for side effects
}
\description{
Same as \code{rm(list=ls())}
}
\examples{
\dontrun{rm_all()}
}
\keyword{live}
