% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{split_docstring}
\alias{split_docstring}
\title{Split a docstring into a Head, Param and Tail Part}
\usage{
split_docstring(docstring)
}
\arguments{
\item{docstring}{Docstring of a R function as string, i.e. as character
vector of length 1.}
}
\value{
List with 3 elements: head, param and tail.
}
\description{
Split a docstring into a head, param and tail part.
}
\examples{
uri <- system.file("testfiles/funcs.R", package = "toscutil")
func <- "f4"
content <- readLines(uri)
docstring <- get_docstring(content, func)
split_docstring(docstring)
}
\keyword{doc}
