% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live.R
\name{stub}
\alias{stub}
\title{Stub Function Arguments}
\usage{
stub(func, ..., envir = parent.frame())
}
\arguments{
\item{func}{function for which the arguments should be stubbed}

\item{...}{non-default arguments of \code{func}}

\item{envir}{environment to which symbols should be assigned}
}
\value{
list of symbols that are assigned to \code{envir}
}
\description{
\code{stub()} assigns all arguments of a given function as symbols
to the specified environment (usually the current environment). For arguments
without default values, \code{stub()} will attempt to retrieve their values from
\code{.GlobalEnv} if they exist there.
}
\details{
Stub is thought to be used for interactive testing and unit testing.
It does not work for primitive functions.

When a function has required arguments without defaults, \code{stub()} will first
check if those arguments exist in \code{.GlobalEnv} and use their values if found.
This enables a common dev workflow: (1) Run example code that sets variables,
(2) Call \code{stub(func)}, (3) Modify and execute parts of the function body.
}
\examples{
f <- function(x, y = 2, z = 3) x + y + z
args <- stub(f, x = 1) # assigns x = 1, y = 2 and z = 3 to current env

# stub() can also use values from GlobalEnv for missing args:
g <- function(a, b = 10) a + b
a <- 5 # Set in GlobalEnv
stub(g) # Uses a = 5 from GlobalEnv, assigns a = 5 and b = 10
}
\keyword{live}
