% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catf.R
\name{catf}
\alias{catf}
\alias{catfn}
\title{Format and Print}
\usage{
catf(
  fmt,
  ...,
  end = "",
  file = "",
  sep = " ",
  fill = FALSE,
  labels = NULL,
  append = FALSE
)

catfn(
  fmt,
  ...,
  end = "\\n",
  file = "",
  sep = " ",
  fill = FALSE,
  labels = NULL,
  append = FALSE
)
}
\arguments{
\item{fmt}{passed on to \code{\link[base:sprintf]{base::sprintf()}}}

\item{...}{passed on to \code{\link[base:sprintf]{base::sprintf()}}}

\item{end}{passed on to \code{\link[=cat2]{cat2()}}}

\item{file}{passed on to \code{\link[=cat2]{cat2()}} (which passes it on to \code{\link[base:cat]{base::cat()}})}

\item{sep}{passed on to \code{\link[=cat2]{cat2()}} (which passes it on to \code{\link[base:cat]{base::cat()}})}

\item{fill}{passed on to \code{\link[=cat2]{cat2()}} (which passes it on to \code{\link[base:cat]{base::cat()}})}

\item{labels}{passed on to \code{\link[=cat2]{cat2()}} (which passes it on to \code{\link[base:cat]{base::cat()}})}

\item{append}{passed on to \code{\link[=cat2]{cat2()}} (which passes it on to \code{\link[base:cat]{base::cat()}})}
}
\value{
No return value, called for side effects
}
\description{
Same as \code{cat2(sprintf(fmt, ...))}
}
\examples{
catf("A\%dB\%sC", 2, "asdf") # prints "A2BasdfC"
catfn("A\%dB\%sC", 2, "asdf") # prints "A2BasdfC\n"
}
