% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{get_pkg_docs}
\alias{get_pkg_docs}
\title{Get Documented Functions in a Package}
\usage{
get_pkg_docs(pkg = "tools")
}
\arguments{
\item{pkg}{The package name.}
}
\value{
Returns a dataframe with columns \code{title}, \code{description}, \code{value}, \code{examples} and rows corresponding to the documented functions in the package.
}
\description{
Lists all documented functions in a package together with some of their documentation elements as raw text.
}
\examples{
df2 <- df <- get_pkg_docs("tools")
for (x in colnames(df2)) df2[[x]] <- substr(df2[[x]], 1, 16)
str(df)
print(head(df2), row.names = FALSE)
}
