% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{help2}
\alias{help2}
\title{Return help for topic}
\usage{
help2(topic, format = "text", package = NULL)
}
\arguments{
\item{topic}{character, the topic for which to return the help text. See argument \code{topic} of function \code{\link[=help]{help()}} for details.}

\item{format}{character, either \code{"text"}, \code{"html"} or \code{"latex"}}

\item{package}{character, the package for which to return the help text. This argument will be ignored if topic is specified. Package must be attached to the search list first, e.g. by calling \code{library(package)}.}
}
\value{
The help text for the specified topic in the specified format as string.
}
\description{
Returns the help text for the specified topic formatted either as plain text, html or latex.
}
\examples{
htm <- help2("sum", "html")
txt <- help2(topic = "sum", format = "text")
cat2(txt)
}
\keyword{base}
