% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{help2}
\alias{help2}
\title{Return help for topic}
\usage{
help2(topic, format = c("text", "html", "latex"), package = NULL)
}
\arguments{
\item{topic}{character, the topic for which to return the help text. See
argument \code{topic} of function \code{\link[=help]{help()}} for details.}

\item{format}{character, either \code{"text"}, \code{"html"} or \code{"latex"}}

\item{package}{character, the package for which to return the help text.
This argument will be ignored if topic is specified. Package must be
attached to the search list first, e.g. by calling \code{library(package)}.}
}
\description{
This function returns the help text for the specified topic
formatted either as plain text, html or latex.
}
\details{
This function was copied in part from:
\url{https://www.r-bloggers.com/2013/06/printing-r-help-files-in-the-console-or-in-knitr-documents/}
}
\examples{
htm <- help2("sum", "html")
txt <- help2(topic = "sum", format = "text")
cat2(txt)
}
\keyword{base}
