% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.none.R
\name{is.none}
\alias{is.none}
\title{Like \href{https://docs.python.org/3/library/functions.html#bool}{pythons \code{bool}}}
\usage{
is.none(x)
}
\arguments{
\item{x}{object to test}
}
\value{
TRUE if \code{x} is FALSE, 0, NULL, NA, an empty list or an empty
string. Else FALSE.
}
\description{
TRUE for FALSE, 0, NULL, NA, empty lists and empty string
}
\examples{
is.none(FALSE) # TRUE
is.none(0) # TRUE
is.none(NA) # TRUE
is.none(list()) # TRUE
is.none("") # TRUE
is.none(1) # FALSE
}
