% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.none.p.R
\name{op-null-default}
\alias{op-null-default}
\alias{\%none\%}
\title{Default operator}
\usage{
x \%none\% y
}
\arguments{
\item{x}{object to test}

\item{y}{object to return if \code{is.none(x)}}
}
\description{
Like rlang's \%||\% but also checks for empty lists and empty strings
}
\examples{
is.none(FALSE) # TRUE
is.none(0) # TRUE
is.none(NA) # TRUE
is.none(list()) # TRUE
is.none("") # TRUEs
is.none(1) # FALSE
}
