% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.none.p.R
\name{op-null-default}
\alias{op-null-default}
\alias{\%none\%}
\title{Default operator}
\usage{
x \%none\% y
}
\arguments{
\item{x}{object to test}

\item{y}{object to return if \code{is.none(x)}}
}
\value{
Returns \code{y} if \code{is.none(x)} else \code{x}
}
\description{
Like rlang's \verb{\%||\%} but also checks for empty lists and empty
strings (for details see
\emph{https://rdrr.io/cran/rlang/man/op-null-default.html}).
}
\examples{
FALSE \%none\% 2 # returns 2
0 \%none\% 2 # returns 2
NA \%none\% 2 # returns 2
list() \%none\% 2 # returns 2
"" \%none\% 2 # returns 2
1 \%none\% 2 # returns 1
}
\seealso{
\code{\link[=is.none]{is.none()}}
}
